Move `EventImageDataLoader` to its own file
